requirements.txt
streamlit
opencv-python-headless
Pillow
easyocr
pytesseract
numpy
pandas
matplotlib
openai
pytest